backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12
openai
edge-tts
httpx==0.27.2
ffmpeg-python==0.2.0
pillow==11.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
